add getContentTypeFix function
This commit adds a getContentTypeFix function to the jselem module. This new function calls getContentType after ensuring the path has a file extension by calling getPath.

The getContentTypeFix function provides an alternative way to get the content type of a file by automatically adding the file extension if missing. This matches the behavior of the other functions in the module that operate on paths.¡